Add a flag in Loop to make it keep rendering
For applications that require lots of animation, this can make the code
a bit simpler since you won’t need to request new frames all over the
place.

This setting should be application specific, and not changed in views.

I might change this to be a private initializer variable later.